% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/asym_alpha.R
\name{asym_alpha}
\alias{asym_alpha}
\title{Asymtotic alpha}
\usage{
asym_alpha(sel, neut, plot = FALSE, main = "", ...)
}
\arguments{
\item{sel}{The unfolded site frequency spectrum for selected sites}

\item{neut}{The unfolded site frequency spectrum for neutral sites}

\item{plot}{Should the points and fitted line be plotted}

\item{main}{Main title passed to plot()}

\item{...}{Additional graphical parameters to pass to plot()}
}
\description{
This function uses the asymptotic approach described by Messer & Petrov (2013)
to estimate the population genetic paramater alpha. This function returns
fittedalpha[0] and alpha[1] as well as an optional figure depicting
alpha[1/n], alpha[2/n],..., alpha[n-1] and the fitted line
}
\examples{
asym_alpha(sel, neut, plot=FALSE, main="", ...)
}
\references{
Messer, P. W., and D. Petrov. 2013. Frequent adaptation and the McDonald–Kreitman
 test. Proceedings of the National Academy of Sciences. 110(21):8615–8620.
}
\keyword{alpha}
\keyword{sfs}

