% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bootstrap_alpha.R
\name{bootstrap_alpha}
\alias{bootstrap_alpha}
\title{Bootstrap alpha}
\usage{
bootstrap_alpha(genes, sel.sfs, neut.sfs, reps = 1000)
}
\arguments{
\item{sel.sfs}{The unfolded site frequency spectrum for selected sites for a colletion of genes one per line}

\item{neut.sfs}{The unfolded site frequency spectrum for neutral sites for a collection of genes one per line}

\item{plot}{Should the points and fitted line be plotted}

\item{main}{Main title passed to plot()}

\item{...}{Additional graphical parameters to pass to plot()}
}
\description{
This function calculates alpha_1 and alpha_0 values for a set of
user specified genes from a large user supplied set, then assesses
the significance of these values by perform a user-specified number
of resampling replicates, drawn from the full set of genes.
Operations are performed by calling the \code{asym_alpha} function.
}
\examples{
data(spectrum2stats)
bootstrap_alpha(genes, sel.sfs, neut.sfs, reps=1000)
}
\references{
Messer, P. W., and D. Petrov. 2013. Frequent adaptation and the McDonald–Kreitman
 test. Proceedings of the National Academy of Sciences. 110(21):8615–8620.
}
\keyword{alpha}
\keyword{bootstrap}
\keyword{sfs}

