% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fold_sfs.R
\name{fold_sfs}
\alias{fold_sfs}
\title{Folded site frequency spectrum}
\usage{
fold_sfs(sfs)
}
\arguments{
\item{sfs}{The unfolded site frequency spectrum}
}
\description{
This function takes an unfolded site frequency spectrum and returns a folded sfs
}
\examples{
fold_sfs(sfs)
}
\keyword{folded_sfs}
\keyword{sfs}

