% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_alpha.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Plot alpha}
\usage{
plot_alpha(df, main = "alpha bootstrap")
}
\arguments{
\item{df}{A dataframe returned by bootstrap_alpha}

\item{main}{An optional main title for each plot}
}
\description{
This function plots the bootstrap distributions returned by boostrap_alpha
}
\examples{
plot_alphadf,main="alpha bootstrap")
}
\keyword{alpha}
\keyword{bootstrap}
\keyword{sfs}

