% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TD_sfs.R
\name{TD_sfs}
\alias{TD_sfs}
\title{Tajima's D}
\usage{
TD_sfs(sfs)
}
\arguments{
\item{sfs}{The unfolded site frequency spectrum}
}
\description{
This function takes an unfolded site frequency spectrum and returns Tajima's D
}
\examples{
TD_sfs(sfs)
}
\keyword{D}
\keyword{TD}
\keyword{Tajima's}
\keyword{sfs}

