% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FWH_sfs.R
\name{FWH_sfs}
\alias{FWH_sfs}
\title{Fay and Wu's H}
\usage{
FWH_sfs(sfs)
}
\arguments{
\item{sfs}{The unfolded site frequency spectrum}
}
\description{
This function takes an unfolded site frequency spectrum and returns Fay and Wu's H
}
\examples{
FWH_sfs(sfs)
}
\references{
Fay, JC and Wu, CI. 2000. Hitchiking under positive Darwinian selection.
Genetics. 155:1405-1413
}
\keyword{FWH}
\keyword{sfs}

