% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mk_sfs.R
\name{mk_sfs}
\alias{mk_sfs}
\title{McDonald-Kreitman test}
\usage{
mk_sfs(sel, neut)
}
\arguments{
\item{sel}{The unfolded site frequency spectrum for selected sites}

\item{neut}{The unfolded site frequency spectrum for neutral sites}
}
\description{
This function performs uses site frequency spectra to create a four
cell contingency table of fixed and segregating, selected and neutral
sites. This table is used to perform a McDonald-Kreitman test, and
caculate summaries (alpha, DoS)
}
\examples{
mk_sfs(sel, neut)
}
\references{
McDonald, J., and M. Kreitman. 1991. Adaptive protein evolution at the Adh locus in Drosophila. Nature 351:652–654.
Eyre-Walker DoS
Smith and Eyre Walker alpha
}
\keyword{alpha}
\keyword{sfs}

